requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
python-magic==0.4.27
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
